requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pyloudnorm>=0.1.1
audiomentations>=0.30.0

# Image Processing
Pillow>=9.5.0
Wand>=0.6.13
opencv-python>=4.7.0
albumentations>=1.3.0

# Pretrained Models
transformers>=4.30.0
diffusers>=0.21.0
timm>=0.9.0
peft>=0.7.0
accelerate>=0.24.0

# Multimodal
open_clip_torch>=2.20.0
laion-clap>=1.1.0
imagebind @ git+https://github.com/facebookresearch/ImageBind.git

# Training & Utils
pytorch-lightning>=2.0.0
wandb>=0.15.0
tensorboard>=2.13.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Metrics & Evaluation
torchmetrics>=1.0.0
mir_eval>=0.7
essentia>=2.1b6.dev1110

# Data & File Handling
h5py>=3.8.0
tqdm>=4.65.0
pandas>=2.0.0
pyyaml>=6.0

# Development
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Audio Generation (specific)
# Note: AudioLDM may need separate installation
# audioldm>=0.0.1  # Install from source if needed
